re-adjusted space b/w nav bar and the home msg
