+++ src/components/landingPage/Hero.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const Hero = () => {
  return (
-   <div className="w-screen h-screen text-white bg-center bg-cover">
+   <div className="w-screen h-screen -my-20 text-white bg-center bg-cover">
      <div class="container mx-auto flex h-full text-center my-auto items-center justify-center flex-col">
        <div class="text-center lg:w-5/12 w-full flex-col">
          <h1 className="mx-auto text-5xl font-bold leading-tight font-kumbh pt-20">
