+++ tailwind.config.js
@@ -9,12 +9,18 @@ export default {
       colors: {
         primaryColor: 'rgba(21, 34, 89, 1)',
         secondaryColor: 'rgba(80, 156, 219, 1)',
-        tertiaryColor: 'rgba(185, 215, 241, 1)'
+        tertiaryColor: 'rgba(185, 215, 241, 1)',
+        btnColor: '#2D88D4',
+        btnHoverColor: '#509CDB',
       },
       fontFamily: {
         'kumbh': ['Kumbh Sans', 'sans-serif']
-      }
+      },
+      backgroundImage: {
+        'hero-bg': "url('./src/assets/hero-bg.png')",
+      },
     },
   },
   plugins: [],
 };
+
